Fix: manager no longer install path dependencies
PathResolver returns the installed spec from the local path (where
the source code is installed) because it links the `src` folder and
thus can't access `shard.yml` from its libs folder.

Yet PathResolver always returned something for the installed spec,
and thus the install and update commands thought it was already
installed, which was wrong.