Drop custom dependency groups; Add --production
Groups don't play well with the lock file. Trying to install custom
groups broke the lock file that no longer contained some shards (eg:
development) but would contain some others (eg: custom).

Only the development group has been kept, along with the ability to
not install this group using the `--production` flag, which can be
useful when deploying an application to be built in production. The
lock file won't be generated or overwritten in this case, and
installation will fail if a dependency has been added but isn't in
the lock file.