Correctly parses empty Projectfile
By default, Crystal, prior to version 0.8.0, used to generate
an empty `Projectfile` with no listed dependencies:

    deps do
    end

Developers might have committed this empty Projectfile to their
repositories, which then was parsed as legacy and attempt to build
the dependencies.

This result in an empty array of dependencies, which was incorrectly
concatenated to the fake spec built by Shards in attempt to process
legacy repositories. This was not proper YAML and thus, blow up
at runtime.

The fix provided here avoids that by resetting the empty list of
dependencies and simply return nothing, generating a proper YAML
and allowing the parsing to continue.

Fixes reported issue #42.