Refactor check command
The check command tried to resolve dependencies just like the
install and update commands, but without cloning repositories. This
approach didn't work for newly cloned projects, which quickly
failed with different errors.

The refactor changes this for loading the spec, checking if each
dependency is installed, or not, and whether the versions match the
requirements and the locks.

This proves to work better, overall.

fixes #51