Fix: install didn't update locked commit dependency
Trying to install dependencies when a locked commit was updated in
the lockfile didn't install the newer version, while claiming that
it did so.

Shards now saves a `.shards/dependency.sha1` to identify the actual
installed commit sha1, so we can correctly guess which fie is
actually installed.

NOTE: this fails for a shared cache directory...

fixes #65 (part 2)