Verify git command line tool presence
Corrects the behavior that a Git command failing was reported as
maybe git isn't installed. Now we check for `git` to be present and
print an error accordingly.

The git command failing error message was changed to reflect that
maybe a branch, commit or file doesn't exist, which is more correct.

closes #108
closes #128