Introduces `version` to retrieve shards current version
Sometimes is useful to obtain the current version string of the
library/project you're working on, either to help you determine if
was bumped (when preparing for a new release) or when integrating
it within the library's `VERSION` constant (through macros).

This change introduces `version` command for such purpose, allowing
you retrieve from current directory (by default) or be able to
provide a different path.

Usage example

    module Foo
      VERSION = {{ `shards version #{__DIR__}`.chomp.stringify }}
    end

Resulting in a `Foo::VERSION` matching the `version` defined in
`shard.yml`.

Fixes #147