Fix: installing Git refs
Specifying a Git refs (tag, branch or commit) used to inject that
ref into the available versions array for the shard. This behavior
caused an issue in NaturalSort when trying to resolve a version
number to install.

The behavior was changed so the Git refs is no longer injected, but
will be reported as the actual "version" to be installed.

Also introduces a `Package#report_version` method which returns a
string describing which version (number) of a shard is installed,
along with the Git refs which is installed. For example:

    Installing dotenv (0.1.0 at master)

fixes #162