Install executables (#126)
Uses (hard) links on POSIX systems to install the executable,
taking care to verify whether the destination isn't already the
source (i.e. same inode). On Windows we just copy the file.

Some filesystems don't support hard links which is reported with
the EPERM errno, or we try to link across mounted file systems.
These cases are edge cases, but we fallback to copy the
executable to avoid errors.