Fix: always resolve dependency spec at locked commit
If the dependency is a branch refs and we previously locked the
dependency, we must declare that locked commit (for the install
command only) otherwise we could end up resolving the spec from the
latest branch commit that may have changed dependencies, while still
eventually installing the locked commit.

This situation is fixed with a hack in Manager to add the locked
commit before the branch dependency is added, so the locked commit
will have precedence over the branch.