Support prereleases
A version number is considered a prerelease when it has at least one
letter. For example `1.0.alpha` and `2.0.0-pre10` are valid
prereleases for release versions `1.0` and `2.0.0` respectively.

Prereleases will never be selected until a `shard.yml` specifically
requires one. For example `version: 1.0-alpha`. The approximate
operator is recommended since it will allow upgrading to the next
prereleases and eventually to the release. For example
`version: ~> 1.0.0.pre` can be updated to `1.0.0.rc` then `1.0.0` but not
`1.1.0.beta`.