Fix: transitive dependencies in postinstall script
Postpones the execution of the postinstall script, so it's run after
all dependencies have been installed, to solve the availability of
transitive dependencies. Also postpones the installation of
executables, since they're likely to be built by the postinstall
script.

Links the project's `lib` folder as the shard's own `lib` folder,
to solve the accessibility of transitive dependencies (for
executable for the postinstall script). This solution was preferred
over manipulating the CRYSTAL_PATH environment variable, since a
wrapper script of the crystal executable could replace it (e.g. the
`bin/script` of crystal's repository does just that).