Add run command
Introduces `shards run`, similar to `shards build`, that builds and runs
the specified target, providing a unified interface to developers.

With this, a developer can directly use `shards run <target>` in a
single step without having first to compile (build) and then execute it
with `./bin/target`. No need for a script or wrapper.

This is inspired by `crystal build` and `crystal run` user experience
(with the difference that `crystal run` creates a temporary binary).

This is very helpful for automatic rebuild scenarios, where a process is
monitoring your code and can rebuild and relaunch your application.

When invoking `shards run`, it will first build the target (with the
supplied options) and then invoke it, passing the extra options:

  $ shards run myapp --stats -- --args

Is the same as:

  $ shards build myapp --stats

  $ ./bin/myapp --args

If the target exits with error, the exit code is passed back.

Resolves #157
Relates #298

Co-authored-by: Elliot DeNolf <65888+denolfe@users.noreply.github.com>