Improve support for older versions of Fossil.
Versions before 2.12 did not have a --workdir argument for the
timeline command, and were more strict about argument ordering.  This
can be faked using Dir#cd and arguments in the correct order.

Versions before 2.14 did not support the --format/-F argument for the
timeline command.  As a workaround, we can use timeline + whatis to
get the full artifact hash.