Fix clear ENV in integration spec (#552)
Clear environment when running integration spec. If the environment has CRYSTAL_OPTS=--error-on-warnings it would get passed through to the compiler and the spec fails (see crystal-lang/test-ecosystem/runs/6116920991).

This is a minimal change to fix the immediate issue. It could be considered that we apply this to more uses of run. Maybe inheriting the environment should even be opt-in.