Fix `GitResolver#valid_repository?` (#646)
`git config` exits with status code 1 if the config value was not found. This would raise an exception but instead, `#valid_repository?` should just return `false`. This patch fixes that. It also moves from `--get-regexp` to plain `--get`. The remote name `origin` is fixed anyway (e.g. `git ls-remote --get-url origin` in `#origin_url`) and we only care about this.